Move drawable file
