+++ app/src/main/java/nil/nadph/qnotified/InjectDelayableHooks.java
@@ -42,9 +42,9 @@ import java.util.Collections;
 import java.util.HashSet;
 import nil.nadph.qnotified.hook.AbsDelayableHook;
 import nil.nadph.qnotified.step.Step;
-import nil.nadph.qnotified.ui.ProportionDrawable;
+import nil.nadph.qnotified.ui.drawable.ProportionDrawable;
 import nil.nadph.qnotified.ui.ResUtils;
-import nil.nadph.qnotified.ui.SimpleBgDrawable;
+import nil.nadph.qnotified.ui.drawable.SimpleBgDrawable;
 import nil.nadph.qnotified.util.LicenseStatus;
 import nil.nadph.qnotified.util.ReflexUtil;
