+++ app/src/main/java/nil/nadph/qnotified/ui/drawable/DebugDrawable.java
@@ -19,7 +19,7 @@
  * <https://www.gnu.org/licenses/>
  * <https://github.com/ferredoxin/QNotified/blob/master/LICENSE.md>.
  */
-package nil.nadph.qnotified.ui;
+package nil.nadph.qnotified.ui.drawable;
 
 import android.content.Context;
 import android.graphics.Canvas;
